# ai-platform/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinOps AI \u2014 platform API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/connections": {
      "get": {
        "summary": "Connections",
        "description": "Every MCP server, its tools, access tier, and enterprise health \u2014 for the portal.",
        "operationId": "connections_connections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/investigate": {
      "post": {
        "summary": "Post Investigate",
        "description": "Run the Investigator: plan -> tool loop -> synthesized Finding, one trace id.",
        "operationId": "post_investigate_investigate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Finding"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trades": {
      "get": {
        "summary": "List Trades",
        "description": "Proxy to the enterprise trade search \u2014 the portal talks only to platform-api.",
        "operationId": "list_trades_trades_get",
        "parameters": [
          {
            "name": "client",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Client"
            }
          },
          {
            "name": "account",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "security",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Security"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TradeRow"
                  },
                  "title": "Response List Trades Trades Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/trades/{trade_id}": {
      "get": {
        "summary": "Get Trade",
        "operationId": "get_trade_trades__trade_id__get",
        "parameters": [
          {
            "name": "trade_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trade Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeRow"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/knowledge": {
      "get": {
        "summary": "Knowledge",
        "description": "Search the SOP / fixture corpus \u2014 for the portal's Knowledge tab.",
        "operationId": "knowledge_knowledge_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Q"
            }
          },
          {
            "name": "k",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 5,
              "title": "K"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KnowledgeHit"
                  },
                  "title": "Response Knowledge Knowledge Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cases": {
      "get": {
        "summary": "List Cases",
        "description": "All cases, newest first \u2014 for the portal's Cases tab.",
        "operationId": "list_cases_cases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "type": "array",
                  "title": "Response List Cases Cases Get"
                }
              }
            }
          }
        }
      }
    },
    "/cases/{case_id}": {
      "get": {
        "summary": "Get Case",
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Case Cases  Case Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/approvals/{approval_id}/decide": {
      "post": {
        "summary": "Decide Approval",
        "description": "A human (role `OPS_ANALYST`) approves or rejects a proposed action.",
        "operationId": "decide_approval_approvals__approval_id__decide_post",
        "parameters": [
          {
            "name": "approval_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Approval Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Decide Approval Approvals  Approval Id  Decide Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConnectionsResponse": {
        "properties": {
          "enterprise_base_url": {
            "type": "string",
            "title": "Enterprise Base Url"
          },
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "servers": {
            "items": {
              "$ref": "#/components/schemas/ServerInfo"
            },
            "type": "array",
            "title": "Servers"
          }
        },
        "type": "object",
        "required": [
          "enterprise_base_url",
          "healthy",
          "servers"
        ],
        "title": "ConnectionsResponse"
      },
      "DecideRequest": {
        "properties": {
          "decision": {
            "type": "string",
            "title": "Decision"
          },
          "decided_by": {
            "type": "string",
            "title": "Decided By"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "default": "OPS_ANALYST"
          }
        },
        "type": "object",
        "required": [
          "decision",
          "decided_by"
        ],
        "title": "DecideRequest"
      },
      "EvidenceRef": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          },
          "cited": {
            "type": "boolean",
            "title": "Cited",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "kind",
          "ref"
        ],
        "title": "EvidenceRef"
      },
      "Finding": {
        "properties": {
          "subject": {
            "$ref": "#/components/schemas/SubjectRef"
          },
          "outcome": {
            "$ref": "#/components/schemas/Outcome"
          },
          "root_cause": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root Cause"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "type": "array",
            "title": "Evidence"
          },
          "proposed_actions": {
            "items": {
              "$ref": "#/components/schemas/ProposedAction"
            },
            "type": "array",
            "title": "Proposed Actions"
          },
          "rejected_alternatives": {
            "items": {
              "$ref": "#/components/schemas/RejectedAlternative"
            },
            "type": "array",
            "title": "Rejected Alternatives"
          },
          "open_questions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Open Questions"
          },
          "checked": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checked"
          },
          "degraded_tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Degraded Tools"
          },
          "confidence_basis": {
            "type": "string",
            "title": "Confidence Basis",
            "default": ""
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id",
            "default": ""
          },
          "case_id": {
            "type": "string",
            "title": "Case Id",
            "default": ""
          },
          "planning_turns": {
            "type": "integer",
            "title": "Planning Turns",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "subject",
          "outcome"
        ],
        "title": "Finding"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvestigateRequest": {
        "properties": {
          "trade_id": {
            "type": "string",
            "title": "Trade Id"
          }
        },
        "type": "object",
        "required": [
          "trade_id"
        ],
        "title": "InvestigateRequest"
      },
      "KnowledgeHit": {
        "properties": {
          "doc": {
            "type": "string",
            "title": "Doc"
          },
          "section": {
            "type": "string",
            "title": "Section",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "doc",
          "text",
          "score"
        ],
        "title": "KnowledgeHit"
      },
      "Outcome": {
        "type": "string",
        "enum": [
          "RESOLVED_CAUSE",
          "INSUFFICIENT_EVIDENCE",
          "TOOL_DEGRADED",
          "OUT_OF_SCOPE",
          "SKELETON"
        ],
        "title": "Outcome"
      },
      "ProposedAction": {
        "properties": {
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "params": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Params"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "impact": {
            "items": {
              "$ref": "#/components/schemas/SubjectRef"
            },
            "type": "array",
            "title": "Impact"
          },
          "reversible": {
            "type": "boolean",
            "title": "Reversible",
            "default": true
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          }
        },
        "type": "object",
        "required": [
          "action_type",
          "rationale"
        ],
        "title": "ProposedAction"
      },
      "RejectedAlternative": {
        "properties": {
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "type": "object",
        "required": [
          "action_type",
          "reason"
        ],
        "title": "RejectedAlternative"
      },
      "ServerInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "access": {
            "type": "string",
            "title": "Access"
          },
          "healthy": {
            "type": "boolean",
            "title": "Healthy"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolInfo"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "type": "object",
        "required": [
          "name",
          "access",
          "healthy",
          "tools"
        ],
        "title": "ServerInfo"
      },
      "SubjectRef": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "id": {
            "type": "string",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id"
        ],
        "title": "SubjectRef"
      },
      "ToolInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "access": {
            "type": "string",
            "title": "Access"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "name",
          "access",
          "description"
        ],
        "title": "ToolInfo"
      },
      "TradeRow": {
        "properties": {
          "trade_id": {
            "type": "string",
            "title": "Trade Id"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "security_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Security Id"
          },
          "qty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Qty"
          },
          "side": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Side"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "trade_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trade Date"
          },
          "settle_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settle Date"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "cpty_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpty Id"
          },
          "booked_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booked At"
          }
        },
        "type": "object",
        "required": [
          "trade_id"
        ],
        "title": "TradeRow"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}